refactor: code refactor and bug fixes (#56)
