+++ src/app/pages/Catalog-page/CatalogPageImages.tsx
@@ -1,0 +1,12 @@
+export const CatalogPageImages: React.FC = () => {
+  return (
+    <>
+      <img src="/images/catalogPageImg/flower-left.png" alt="flower" className="absolute z-[2]" />
+      <img src="/images/catalogPageImg/flower-right.png" alt="flower" className="absolute right-0 z-[2]" />
+      <img src="/images/catalogPageImg/ellipse-l1.png" alt="ellipse" className="absolute top-[100px] z-[1]" />
+      <img src="/images/catalogPageImg/ellipse-l2.png" alt="ellipse" className="absolute top-[550px] z-[1]" />
+      <img src="/images/catalogPageImg/ellipse-l3.png" alt="ellipse" className="absolute top-[750px] z-[1]" />
+      <img src="/images/catalogPageImg/elipse-r-t.png" alt="ellipse" className="absolute right-0 z-[1]" />
+    </>
+  )
+}
