+++ src/app/pages/Login-page/login-Images.tsx
@@ -4,43 +4,43 @@ export const LoginImages: React.FC = () => {
       <img
         className="absolute right-0 top-0 z-10 w-1/4 object-cover"
         src="/images/loginPageImg/loginPage-flower1.png"
-        alt=""
+        alt="loginPage-flower"
         data-testid="loginFlower1"
       />
       <img
         className="absolute bottom-0 left-0 z-10 w-1/4 object-cover"
         src="/images/loginPageImg/loginPage-flower2.png"
-        alt=""
+        alt="loginPage-flower"
         data-testid="loginFlower2"
       />
       <img
         className=" z-5 absolute right-0 top-0  w-1/4 object-cover opacity-60"
         src="/images/loginPageImg/Ellipse rt.png"
-        alt=""
+        alt="loginPage-flower"
         data-testid="loginEllipseRt"
       />
       <img
         className=" z-5 absolute bottom-0 left-0  w-1/4 object-cover opacity-60"
         src="/images/loginPageImg/Ellipse lb.png"
-        alt=""
+        alt="loginPage-flower"
         data-testid="loginEllipseLb"
       />
       <img
         className=" z-5 absolute left-0 top-0  w-2/12 object-cover opacity-60"
         src="/images/loginPageImg/Ellipse lt.png"
-        alt=""
+        alt="loginPage-flower"
         data-testid="loginEllipseLt"
       />
       <img
         className=" z-5 absolute bottom-0 right-0  w-2/12 object-cover opacity-60"
         src="/images/loginPageImg/Ellipse rb.png"
-        alt=""
+        alt="loginPage-flower"
         data-testid="loginEllipseRb"
       />
       <img
         className=" z-5 absolute left-1/2 top-1/2 w-[250px]  object-cover "
         src="/images/loginPageImg/signature.png"
-        alt=""
+        alt="loginPage-flower"
         data-testid="loginSignature"
       />
     </>
