+++ src/app/pages/NotFound-page/NotFound-page.tsx
@@ -1,5 +1,7 @@
 import { NavLink } from 'react-router-dom'
 
+import { NotFoundImages } from './NotFoundImages'
+
 export const NotFoundPage: React.FC = () => {
   return (
     <div className="relative top-0  flex w-[100%] flex-grow flex-col  overflow-hidden bg-black">
@@ -17,19 +19,7 @@ export const NotFoundPage: React.FC = () => {
           MAIN PAGE
         </NavLink>
       </div>
-      <img
-        src="/notFound/signaturePink.png"
-        alt="signaturePink"
-        className="z-10 mt-[30px] hidden h-[140px] w-[200px] sm:ml-[25%] lg:block"
-      />
-      <img src="/notFound/pinkImg.png" alt="mainImg" className="absolute bottom-0 right-0 z-[5] sm:bottom-[-100px]" />
-      <img src="/notFound/eclipsePink.png" alt="ellipsePink" className="absolute  top-[350px]" />
-      <img src="/notFound/bigEllipsePink.png" alt="biEllipsePink" className="absolute  right-0  blur-[200px]" />
-      <img
-        src="/notFound/ellipseTurquoise.png"
-        alt="ellipse Turquoise"
-        className="absolute top-[0]  rotate-[21.1deg] bg-turquoiseEllipse blur-[250px] sm:bottom-[-750px]  sm:right-[-150px] sm:top-[100%] "
-      />
+      <NotFoundImages />
     </div>
   )
 }
