+++ src/app/types/types.ts
@@ -40,4 +40,32 @@ interface IProduct {
   discountPrice?: number
 }
 
+export const enum Constants {
+  MESSAGE_SHOW_TIME = 2000,
+}
+
+export interface ICollectedAddress {
+  country: string | undefined
+  city: string | undefined
+  streetName: string | undefined
+  postalCode: string | undefined
+}
+
+export interface ICollectedAddressField {
+  id: string | undefined
+  radioOption: string | null
+  address: ICollectedAddress
+}
+
+export interface IProfileModalMessage {
+  isShowMessage: boolean
+  text: string
+}
+
+export interface IMainInfoObject {
+  email?: string
+  firstName?: string
+  lastName?: string
+  dateOfBirth?: string
+}
 export type { IFields, OnDataSend, IproductData, IproductInfo, IProduct }
